requirements.txt
absl-py==0.9.0
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
astunparse==1.6.3
atomicwrites==1.4.0
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
blis==0.4.1
boto==2.49.0
boto3==1.15.5
botocore==1.18.5
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
cymem==2.0.3
Cython==0.29.14
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.0.7
djangorestframework==3.11.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
funcy==1.15
future==0.18.2
gast==0.3.3
gensim==3.8.3
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==1.0.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
iniconfig==1.0.1
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.5
jupyter-core==4.6.3
jupyter-http-over-ws==0.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
murmurhash==1.0.2
mysql==0.0.2
mysql-connector-python==8.0.23
mysqlclient==2.0.3
nbconvert==5.6.1
nbformat==5.0.7
nltk==3.5
notebook==6.0.3
numexpr==2.7.1
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.0
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.1.2
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.13.0
psycopg2==2.8.5
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
Pygments==2.6.1
pyLDAvis==2.1.2
pylint==2.5.3
pymongo==3.11.2
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==6.1.0
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.17.0
pytz==2020.1
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.11.0
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
smart-open==2.1.1
spacy==2.3.2
SQLAlchemy==1.4.5
sqlparse==0.3.1
srsly==1.0.2
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thinc==7.4.1
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
Unidecode==1.2.0
urllib3==1.25.10
virtualenv==20.0.25
wasabi==0.8.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wordcloud==1.8.0
wrapt==1.12.1
WTForms==2.3.3
